# cmd/api/docs/swagger.yaml
definitions:
...
  request.Update:
    properties:
      order_id:
        type: integer
      status_id:
        type: integer
    required:
    - order_id
    - status_id
    type: object
...
  respondse.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/order/Allorders:
    get:
      consumes:
      - application/json
      operationId: FindAllordersInshop
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: FindAllorders In admin side
      tags:
      - AdminOrder
  /admin/order/Status:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all orderStatuses
      operationId: List-all-orderStatus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: for geting all order status list
      tags:
      - AdminOrder
  /admin/order/UpdateStatus:
    patch:
      consumes:
      - application/json
      description: update the order statuses by every orderid.
      operationId: Order_status
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Updateorderstatus
      tags:
      - AdminOrder
...
  /order/cancel/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for cancelling an order associated with a user
      operationId: cancel-order
      parameters:
      - description: ID of the order to be cancelled
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Cancels a specific order for the currently logged in user
      tags:
      - UserOrder
  /order/listall:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all orders
      operationId: List-all-orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: for geting all order list
      tags:
      - UserOrder
  /order/place_order/{payment_id}:
    post:
      consumes:
      - application/json
      description: This endpoint allows a user to purchase all items in their cart
      operationId: buyAll
      parameters:
      - description: payment_id
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Buy all items from the user's cart
      tags:
      - Cart
  /order/return/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for Returning an order associated with a user
      operationId: return-order
      parameters:
      - description: ID of the order to be Returned
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Return a specific order for the currently logged in user
      tags:
      - UserOrder
  /order/view/{order_id}:
    get:
      consumes:
      - application/json
      description: retrieving the details of a specific order identified by its order
        ID.
      operationId: view-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched order details
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Failed to fetch order details
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: to get order details
      tags:
      - UserOrder
...
swagger: "2.0"